+++ app/controllers/manage_accounts_controller.rb
@@ -1,6 +1,6 @@
 class ManageAccountsController < ApplicationController
     def index
-        @admins = Admin.paginate(page: params[:admins_page], per_page: 1)
-        @staffs = Staff.paginate(page: params[:staffs_page], per_page: 1)
+        @admins = Admin.paginate(page: params[:admins_page], per_page: 5)
+        @staffs = Staff.paginate(page: params[:staffs_page], per_page: 5)
     end
 end